The standardization update: Some random easy bug fixes: - Fixed a problem with a link to the section csv-txt of the README files. - Minor adjusts in the wording of the documentation. - Fixed a bug in the dock light where it would be a little out of position because of the padding and margins from the parents of the panel-region element. Now it is using the offSetTop, offSetLeft, offSetWidth and offSetHeight to get the right position relative to the positioned panel-region element.
Most of the time was spent in making the program follow the format of
the transformations presented in the documentation, and also to update
the documentation to allow for more formats and flexibility:
- Created a transformations.js file to handle the cleaning of the
transformation files loaded instead of letting the GraphK object handle
it in setTransforms. It is also now more complex, before no checks
would be made to verify the validity of the values loaded, now all
values are checked. I tried to do a thorough test for all cases and it
seems to be working very well.
- The transformations.js file also returns an array of errors found in
the values, at the moment i am just ignoring this, but hopefully it can
be used in the future to help users identify and correct mistakes.
- Changed the transformations documentation:
  - Put all non optional keys at the top of the reference.
  - Changed how the 'normal' value for `type` key works.
  - Removed the key `pkgName`, it was substituted by the `name` key
  already being used to name single transforms.
  - Added two keys: `input` and `outputs`. This is initially planned to
  be used in the routineMaker to match the input and output plugs based
  on type. Also, the `outputs` key allow for defining multiple outputs.
  - Added a section to explain better the `func` key.

The program does not still suport all the keys as defined in the
documentation, i planned to do that in this update but i found it better
to split into: defining the format and being able to load the transforms
properly with the transforms.js file; and then make the program respond
correctly to the keys given. Hope i can do that in this next update.